refactor: Rename React folders inside pages to start with _
